refactor: make playback controls widget public
